CGFloat -> float for qiniu sdk, error occured compiled in xcode6

+++ JLWeChat/Utilities/QNResourceManager.h
@@ -18,7 +18,7 @@
 
 //upload file with urlkey
 - (void)uploadFileWithUrlkey:(NSString *)urlkey
-               progressBlock:(void (^)(NSString *key, CGFloat progress))progressBlock
+               progressBlock:(void (^)(NSString *key, float progress))progressBlock
                completeBlock:(void (^)(BOOL success,  NSString *key))completeBlock;
 
 // download file with url
