+++ JLWeChat/Utilities/QNResourceManager.m
@@ -60,7 +60,7 @@
 #pragma mark - Upload & Download File (Audio & Video)
 
 - (void)uploadFileWithUrlkey:(NSString *)urlkey
-               progressBlock:(void (^)(NSString *key, CGFloat progress))progressBlock
+               progressBlock:(void (^)(NSString *key, float progress))progressBlock
                completeBlock:(void (^)(BOOL success,  NSString *key))completeBlock
 {
     NSString *token = [QNAuthPolicy defaultToken];
